# vips_thumbnail from file and optimizations