--- a/examples/thumbnail/bench_test.go
+++ b/examples/thumbnail/bench_test.go
@@ -2,7 +2,7 @@ package thumbnail
 
 import (
 	"fmt"
-	"github.com/cshum/govips/v2/vips"
+	"github.com/davidbyttow/govips/v2/vips"
 	"io/ioutil"
 	"testing"
 )