--- a/vips/resample.c
+++ b/vips/resample.c
@@ -26,15 +26,15 @@ int resize_image(VipsImage *in, VipsImage **out, double scale, gdouble vscale,
 }
 
 int thumbnail_image(VipsImage *in, VipsImage **out, int width, int height,
-                    int crop) {
+                    int crop, int size) {
   return vips_thumbnail_image(in, out, width, "height", height, "crop", crop,
-                              NULL);
+                              "size", size, NULL);
 }
 
 int thumbnail_buffer(void *buf, size_t len, VipsImage **out,
-                    int width, int height, int crop) {
+                    int width, int height, int crop, int size) {
   return vips_thumbnail_buffer(buf, len, out, width, "height", height,
-                              "crop", crop, NULL);
+                              "crop", crop, "size", size, NULL);
 }
 
 int mapim(VipsImage *in, VipsImage **out, VipsImage *index) {