--- a/vips/resample.c
+++ b/vips/resample.c
@@ -25,6 +25,12 @@ int resize_image(VipsImage *in, VipsImage **out, double scale, gdouble vscale,
   return vips_resize(in, out, scale, "kernel", kernel, NULL);
 }
 
+int thumbnail(const char *filename, VipsImage **out,
+                    int width, int height, int crop, int size) {
+  return vips_thumbnail(filename, out, width, "height", height,
+                              "crop", crop, "size", size, NULL);
+}
+
 int thumbnail_image(VipsImage *in, VipsImage **out, int width, int height,
                     int crop, int size) {
   return vips_thumbnail_image(in, out, width, "height", height, "crop", crop,