--- a/vips/resample.go
+++ b/vips/resample.go
@@ -22,6 +22,17 @@ const (
 	KernelMitchell Kernel = C.VIPS_KERNEL_MITCHELL
 )
 
+// Size represents VipsSize type
+type Size int
+
+const (
+	SizeBoth  Size = C.VIPS_SIZE_BOTH
+	SizeUp    Size = C.VIPS_SIZE_UP
+	SizeDown  Size = C.VIPS_SIZE_DOWN
+	SizeForce Size = C.VIPS_SIZE_FORCE
+	SizeLast  Size = C.VIPS_SIZE_LAST
+)
+
 // https://libvips.github.io/libvips/API/current/libvips-resample.html#vips-resize
 func vipsResize(in *C.VipsImage, scale float64, kernel Kernel) (*C.VipsImage, error) {
 	incOpCounter("resize")
@@ -51,11 +62,11 @@ func vipsResizeWithVScale(in *C.VipsImage, scale, vscale float64, kernel Kernel)
 	return out, nil
 }
 
-func vipsThumbnail(in *C.VipsImage, width, height int, crop Interesting) (*C.VipsImage, error) {
+func vipsThumbnail(in *C.VipsImage, width, height int, crop Interesting, size Size) (*C.VipsImage, error) {
 	incOpCounter("thumbnail")
 	var out *C.VipsImage
 
-	if err := C.thumbnail_image(in, &out, C.int(width), C.int(height), C.int(crop)); err != 0 {
+	if err := C.thumbnail_image(in, &out, C.int(width), C.int(height), C.int(crop), C.int(size)); err != 0 {
 		return nil, handleImageError(out)
 	}
 
@@ -63,7 +74,7 @@ func vipsThumbnail(in *C.VipsImage, width, height int, crop Interesting) (*C.Vip
 }
 
 // https://www.libvips.org/API/current/libvips-resample.html#vips-thumbnail-buffer
-func vipsThumbnailFromBuffer(buf []byte, width, height int, crop Interesting) (*C.VipsImage, ImageType, error) {
+func vipsThumbnailFromBuffer(buf []byte, width, height int, crop Interesting, size Size) (*C.VipsImage, ImageType, error) {
 	src := buf
 	// Reference src here so it's not garbage collected during image initialization.
 	defer runtime.KeepAlive(src)
@@ -87,7 +98,7 @@ func vipsThumbnailFromBuffer(buf []byte, width, height int, crop Interesting) (*
 
 	var out *C.VipsImage
 
-	if err := C.thumbnail_buffer(unsafe.Pointer(&src[0]), C.size_t(len(src)), &out, C.int(width), C.int(height), C.int(crop)); err != 0 {
+	if err := C.thumbnail_buffer(unsafe.Pointer(&src[0]), C.size_t(len(src)), &out, C.int(width), C.int(height), C.int(crop), C.int(size)); err != 0 {
 		return nil, ImageTypeUnknown, handleImageError(out)
 	}