--- a/vips/resample.h
+++ b/vips/resample.h
@@ -12,8 +12,8 @@ int affine_image(VipsImage *in, VipsImage **out, double a, double b, double c,
 int resize_image(VipsImage *in, VipsImage **out, double scale, gdouble vscale,
                  int kernel);
 int thumbnail_image(VipsImage *in, VipsImage **out, int width, int height,
-                    int crop);
+                    int crop, int size);
 int thumbnail_buffer(void *buf, size_t len, VipsImage **out, int width, int height,
-                    int crop);
+                    int crop, int size);
 int mapim(VipsImage *in, VipsImage **out, VipsImage *index);
 int maplut(VipsImage *in, VipsImage **out, VipsImage *lut);