--- a/vips/resample.h
+++ b/vips/resample.h
@@ -11,6 +11,8 @@ int affine_image(VipsImage *in, VipsImage **out, double a, double b, double c,
                  double d, VipsInterpolate *interpolator);
 int resize_image(VipsImage *in, VipsImage **out, double scale, gdouble vscale,
                  int kernel);
+int thumbnail(const char *filename, VipsImage **out, int width, int height,
+                    int crop, int size);
 int thumbnail_image(VipsImage *in, VipsImage **out, int width, int height,
                     int crop, int size);
 int thumbnail_buffer(void *buf, size_t len, VipsImage **out, int width, int height,