… resources/bmp.TestThumbnail_Decode_BMP-macos-11.golden.jpeg
… resources/jpg-8bit-grey-icc-dot-gain.TestThumbnail_NoUpscale-macos-11.golden.jpeg
… resources/png-24bit.TestThumbnail_PNG_CropCentered-macos-11.golden.png
… vips/image_golden_test.go