`NewThumbnailFromFile` and `NewThumbnailFromBuffer` uses [vips_thumbnail](https://www.libvips.org/API/current/libvips-resample.html#vips-thumbnail) and [vips_thumbnail_buffer](https://www.libvips.org/API/current/libvips-resample.html#vips-thumbnail-buffer) with code optimizations, are performing over 2x faster than the same operations using `NewImageFromFile` and `NewImageFromBuffer`

`NewThumbnailFromFile` is slightly slower than `NewThumbnailFromBuffer` but using much less memory, due to Go bytes not needed to be loaded. 

Also added function `vipsDetermineImageTypeFromMetaLoader`, which determine the image type by the [VIPS-META-LOADER](https://www.libvips.org/API/current/libvips-header.html#VIPS-META-LOADER:CAPS) without needing to load the Go bytes.

Benchmark script
```
% go test -bench=. -benchtime=30s -benchmem ./examples/thumbnail
```
result:
```
 go test -bench=. -benchtime=30s -benchmem ./examples/thumbnail
goos: darwin
goarch: amd64
pkg: github.com/davidbyttow/govips/v2/examples/thumbnail
cpu: Intel(R) Core(TM) i5-7360U CPU @ 2.30GHz
BenchmarkNewImageFromFile-4                   87         425604727 ns/op         2600992 B/op        236 allocs/op
BenchmarkNewImageFromBuffer-4                 84         406739430 ns/op         2600351 B/op        225 allocs/op
BenchmarkNewThumbnailFromFile-4              186         191294505 ns/op          261270 B/op        190 allocs/op
BenchmarkNewThumbnailFromBuffer-4            218         165247179 ns/op         2598531 B/op        224 allocs/op
PASS
ok      github.com/davidbyttow/govips/v2/examples/thumbnail   180.474s
```

Benchmark functions consist of thumbnail + flip operations:
```go
func BenchmarkNewImageFromFile(b *testing.B) {
	resizeToTest := func(size int) {
		image, err := vips.NewImageFromFile(file)
		if err != nil {
			panic(err)
		}
		if err := image.Thumbnail(size*100, size*100, vips.InterestingCentre); err != nil {
			panic(err)
		}
		if err := image.Flip(vips.DirectionVertical); err != nil {
			panic(err)
		}
		if _, _, err = image.ExportNative(); err != nil {
			panic(err)
		}
	}
	for i := 0; i < b.N; i++ {
		resizeToTest(1)
		resizeToTest(2)
		resizeToTest(3)
		resizeToTest(4)
		resizeToTest(5)
	}
}

func BenchmarkNewImageFromBuffer(b *testing.B) {
	resizeToTest := func(size int) {
		buf, err := ioutil.ReadFile(file)
		if err != nil {
			panic(err)
		}
		image, err := vips.NewImageFromBuffer(buf)
		if err != nil {
			panic(err)
		}
		if err := image.Thumbnail(size*100, size*100, vips.InterestingCentre); err != nil {
			panic(err)
		}
		if err := image.Flip(vips.DirectionVertical); err != nil {
			panic(err)
		}
		if _, _, err = image.ExportNative(); err != nil {
			panic(err)
		}
	}
	for i := 0; i < b.N; i++ {
		resizeToTest(1)
		resizeToTest(2)
		resizeToTest(3)
		resizeToTest(4)
		resizeToTest(5)
	}
}

func BenchmarkNewThumbnailFromFile(b *testing.B) {
	resizeToTest := func(size int) {
		image, err := vips.NewThumbnailFromFile(file, size*100, size*100, vips.InterestingCentre)
		if err != nil {
			panic(err)
		}
		if err := image.Flip(vips.DirectionVertical); err != nil {
			panic(err)
		}
		if _, _, err = image.ExportNative(); err != nil {
			panic(err)
		}
	}
	for i := 0; i < b.N; i++ {
		resizeToTest(1)
		resizeToTest(2)
		resizeToTest(3)
		resizeToTest(4)
		resizeToTest(5)
	}
}

func BenchmarkNewThumbnailFromBuffer(b *testing.B) {
	resizeToTest := func(size int) {
		buf, err := ioutil.ReadFile(file)
		if err != nil {
			panic(err)
		}
		image, err := vips.NewThumbnailFromBuffer(buf, size*100, size*100, vips.InterestingCentre)
		if err != nil {
			panic(err)
		}
		if err := image.Flip(vips.DirectionVertical); err != nil {
			panic(err)
		}
		if _, _, err = image.ExportNative(); err != nil {
			panic(err)
		}
	}
	for i := 0; i < b.N; i++ {
		resizeToTest(1)
		resizeToTest(2)
		resizeToTest(3)
		resizeToTest(4)
		resizeToTest(5)
	}
}
```